cleanup, add comments, fix release builds
