+++ packages/expo-dev-menu/android/src/release/java/expo/modules/devmenu/DevMenuPackageDelegate.kt
@@ -1,0 +1,8 @@
+package expo.modules.devmenu
+
+import android.content.Context
+import expo.modules.core.interfaces.ReactActivityHandler
+
+object DevMenuPackageDelegate {
+  fun createReactActivityHandlers(activityContext: Context?): List<ReactActivityHandler> = emptyList()
+}
