+++ packages/expo-modules-core/android/src/main/java/expo/modules/core/interfaces/ReactActivityDelegateHandler.java
@@ -9,7 +9,7 @@ import androidx.annotation.Nullable;
 
 /**
  * A handler API for modules to override default ReactActivityDelegate behaviors.
- * Used by {@link ReactActivityDelegateWrapper}
+ * Used by {@link expo.modules.ReactActivityDelegateWrapper}
  */
 public interface ReactActivityDelegateHandler {
   /**
@@ -21,9 +21,18 @@ public interface ReactActivityDelegateHandler {
     return null;
   }
 
+  /**
+   * Called once, at the time the wrapper object is initialized. Opportunity for a module to declare
+   * that this wrapper should ignore certain method calls and not pass them through to the wrapped
+   * delegate.
+   * @return true if this wrapper should no-op on certain method calls, false otherwise
+   */
   default boolean shouldNoop() {
     return false;
   }
 
+  /**
+   * Called when the wrapper object is initialized.
+   */
   default void onWillCreateReactActivityDelegate(ReactActivity activity) {}
 }
